fix: correct generated types + some tests
