+++ global.d.ts
@@ -1,3 +1,5 @@
+/// <reference types="@emotion/react/types/css-prop" />
+
 declare module '*.svg' {
   const content: string
   export default content
