+++ src/components/Expandable/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import PropTypes from 'prop-types'
 import React, { FunctionComponent, ReactNode } from 'react'
-import Box from '../Box'
+import Box, { XStyledProps } from '../Box'
 
 type ExpandableProps = {
   /**
