+++ src/components/Icon/index.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled'
 import PropTypes from 'prop-types'
 import React, { Ref, forwardRef, useMemo } from 'react'
 import { Color } from '../../theme/colors'
-import Box from '../Box'
+import Box, { XStyledProps } from '../Box'
 
 // Non Material Design icons: 'send',
